"Updated form action attribute to point to new URL"
This commit message accurately describes the code diff, which shows that the "action" attribute of the form element in the HTML code has been updated to point to a new URL (<http://77.37.86.215/site2/submit>). The previous URL (/) has been replaced with the new one. The rest of the code remains unchanged.